# interface/api.yml
openapi: 3.0.3
info:
  title: Cambridge API
  description: A REST API for Cambridge application
  version: 1.0.0
  contact:
    name: API Support
    email: support@cambridge.com

servers:
  - url: http://localhost:8000
    description: Development server

paths:
  /health:
    get:
      summary: Health check endpoint
      description: Returns the health status of the API
      operationId: getHealth
      responses:
        '200':
          description: API is healthy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Internal server error"

  /api/v1/add_case:
    post:
      summary: Add a new case with user prompt analysis
      description: Takes a user prompt and returns arguments with related legal cases
      operationId: addCase
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddCaseRequest'
      responses:
        '200':
          description: Case analysis completed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalysisResponse'
        '400':
          description: Bad request - invalid user prompt
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/v1/gen_draft:
    post:
      summary: Generate a legal draft for a case
      description: Takes a case ID and returns a generated legal draft text
      operationId: genDraft
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenDraftRequest'
      responses:
        '200':
          description: Draft generated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenDraftResponse'
        '404':
          description: Case not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

components:
  schemas:
    AddCaseRequest:
      type: object
      properties:
        user_prompt:
          type: string
          description: The user prompt to analyze
          example: "Can I sue my employer for wrongful termination?"
      required:
        - user_prompt
    AnalysisResponse:
      type: object
      properties:
        caseId:
          type: string
          description: Unique identifier for the analyzed case
          example: "CASE-2024-001"
        strengths:
          type: array
          items:
            $ref: '#/components/schemas/Argument'
        weaknesses:
          type: array
          items:
            $ref: '#/components/schemas/Argument'
    
    Argument:
      type: object
      properties:
        argument:
          type: string
          description: The legal argument
          example: "Wrongful termination based on discrimination"
        case_references:
          type: array
          items:
            $ref: '#/components/schemas/CaseReference'
    
    CaseReference:
      type: object
      properties:
        caseIdentifier:
          type: string
          description: Unique identifier for the case
          example: "CASE-2024-001"
        title:
          type: string
          description: Title of the case
          example: "Smith v. ABC Corporation"
        Date:
          type: string
          format: date
          description: Date of the case
          example: "2024-01-15"
          nullable: true
        matchingDegree:
          type: number
          format: double
          description: Matching degree score
          minimum: 0.0
          maximum: 1.0
          example: 0.85
        fileReference:
          type: string
          description: Reference to the case file
          example: "doc_12345.pdf"
    
    HealthResponse:
      type: object
      properties:
        status:
          type: string
          example: "healthy"
        timestamp:
          type: string
          format: date-time
          example: "2024-01-01T12:00:00Z"
        version:
          type: string
          example: "1.0.0"
    
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
        code:
          type: string
          description: Error code

    GenDraftRequest:
      type: object
      properties:
        case_id:
          type: string
          description: The ID of the case to generate a draft for
          example: "CASE-2024-001"
      required:
        - case_id
    
    GenDraftResponse:
      type: object
      properties:
        text:
          type: string
          description: The generated legal draft text
          example: "DRAFT LEGAL DOCUMENT..."
      required:
        - text 